requirements.txt
PyOpenGL
PyGLM
PyGLFW
glfw
numpy
loguru
moderngl
moderngl_window
PyOpenGL_accelerate
freetype-py
